декстоп рантайм обновлянтус

+++ catroid/src/main/java/org/catrobat/catroid/content/actions/DeleteFolderAction.kt
@@ -17,7 +17,12 @@ class DeleteFolderAction : TemporalAction() {
 
         try {
             val downloadsDir = File(RuntimeServicesHolder.services.getDownloadsDir())
-            val folder = File(downloadsDir, name)
+            val baseCanonical = downloadsDir.canonicalPath
+            val folder = File(downloadsDir, name).canonicalFile
+            if (!folder.canonicalPath.startsWith(baseCanonical + File.separator) && folder.canonicalPath != baseCanonical) {
+                Log.e("DeleteFolderAction", "Path traversal detected: $name")
+                return
+            }
             if (folder.exists() && folder.isDirectory) {
                 folder.deleteRecursively()
             }
